Update create_keys.sh script to improve security.
This removes the -E flag as it would break builds, and also downgraded
the hash algorithm from the default sha256 to md5. Additionally, this
changes the ECDSA (NIST NSA curve) keytype to Ed25519 and sets the key
size to 512 bits. For RSA the key size has been increased from the
default to 4096 bits.